API Shift TestAssetStore to non-test dev namespace to ensure this class is available for other modules
